+++ modules/raw-materials/roll-web/src/test/groovy/io/github/springroll/web/request/HandlerHolderTest.groovy
@@ -1,4 +1,4 @@
-package io.github.springroll.web
+package io.github.springroll.web.request
 
 import io.github.springroll.test.AbstractSpringTest
 import org.junit.Test
